refactoring to include tensorflow imports if available
