+++ tcn/hollow_class/__init__.py
@@ -1,0 +1,1 @@
+from tcn.hollow_class import hollow_class
