+++ tcn/keras/__init__.py
@@ -1,0 +1,1 @@
+from .tcn import TCN, compiled_tcn
